Filter in '/store/unmerged' the same way we used to.
